+++ src/parser/enrichers/monster/SpiritualWeapon/_module.mjs
@@ -1,0 +1,1 @@
+export { default as MoveAndAttack } from "./Attack.mjs";