+++ src/parser/enrichers/monster/_module.mjs
@@ -11,3 +11,4 @@ export * as SummonConstruct from "./SummonConstruct/_module.mjs";
 export * as SummonAnimatedObject from "./SummonAnimatedObject/_module.mjs";
 export * as ElementalCultist from "./ElementalCultist/_module.mjs";
 export * as Empyrean from "./Empyrean/_module.mjs";
+export * as SpiritualWeapon from "./SpiritualWeapon/_module.mjs";